Modify install scripts and README
Source ~/.bashrc doesn't usually work from the script
as it is not interactive. Split the installation scripts
so users are requested to do it before the second script.

Updated README by moving common sections out of the
installation and better structuring the logical flow.

Removed Manual file as data ported to Wiki pages.

Signed-off-by: Niranjhana Narayanan <nniranjhana@ece.ubc.ca>